AChoirX v10.01.78c
Added OSQuery to Windows, Linux, and macOS toolkits and scripts (commented out).  This dramatically increases the size of the collectors - Remove it if you don't want it.

This is in preparation for some additional work I am going to do on collection at-scale (parsing targeted artifacts at the endpoint to reduce the collection size when collecting at-scale)